Cleaned CSS and optinput up a little.
