--- Dockerfile
@@ -1,20 +1,0 @@
-# Use the official Node.js image as the base image
-FROM node
-
-# Set the working directory
-WORKDIR /app
-
-# Copy package.json and package-lock.json to the working directory
-COPY package*.json ./
-
-# Install the dependencies
-RUN npm install
-
-# Copy the rest of the application code to the working directory
-COPY . .
-
-# Expose the port the app runs on
-EXPOSE 5173
-
-# Command to run the application
-CMD ["npm", "run", "dev", "--", "--host"]
