--- src/routes/game/components/elements/playerInputPopUp/components/ReorderOpt.module.css
@@ -1,268 +1,0 @@
-.reorderCards {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-  flex-grow: 1;
-  gap: 0.5em;
-  padding: 0 0.5em;
-  height: 50%;
-  width: 100%;
-  margin: 0;
-}
-
-.reorderItem {
-  height: var(--card-size);
-  width: var(--card-size);
-  margin: 0;
-}
-
-.reorderItem {
-  max-width: var(--card-size);
-  color: rgba(0, 0, 0, 0);
-}
-
-.reorderItem ::marker {
-  visibility: none;
-}
-
-.countersCover {
-  position: absolute;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-  inset: 0;
-  z-index: 2;
-  cursor: default;
-}
-
-.number {
-  border-radius: 100%;
-  border-style: solid;
-  border-color: black;
-  border-width: 2px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 2em;
-  width: 2em;
-  font-size: 1.5em;
-  background-color: var(--white);
-  text-align: center;
-  cursor: default;
-}
-
-.text {
-  color: black;
-}
-
-.subtitle {
-  font-size: 1.25em;
-  padding: 0 0 0 0.25em;
-  margin: 0px;
-}
-
-.orderingExplanation {
-  font-size: 0.75em;
-  padding-left: 0.5em;
-  margin-bottom: 0;
-}
-
-.optionsContainer {
-  position: fixed;
-  display: flex;
-  flex-direction: column;
-  min-width: 30%;
-  max-width: 90%;
-  top: 60%;
-  bottom: 5vh;
-  left: 15%;
-  right: 25%;
-  background-color: var(--near-black);
-  padding: 1.5vh;
-  border: 1px solid var(--primary);
-  border-radius: 20px;
-  z-index: 900;
-}
-
-.optionsTitleContainer {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  flex-grow: 0;
-  width: 100%;
-  justify-content: space-between;
-  align-items: center;
-  padding-left: 1vh;
-  padding-bottom: 1vh;
-}
-
-.title {
-  font-size: 18px;
-  padding: 0px;
-  margin: 0px;
-}
-
-.subtitle {
-  font-size: 0.7em;
-  padding: 0px;
-  margin: 2px;
-  font-weight: normal;
-}
-
-.buttonList {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  padding-left: 0.45em;
-  gap: 1.2dvw;
-}
-
-.buttonDiv {
-  background-color: var(--primary);
-  color: black;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  min-width: calc(6dvw * var(--card-scale));
-  height: 32px;
-  border-radius: 5px;
-  border: 1px solid var(--primary);
-  flex-basis: auto;
-  padding: 8px;
-  width: auto;
-  margin-bottom: 0.25rem;
-}
-
-.buttonDiv:hover {
-  background-color: black;
-  color: var(--primary);
-  border: 1px solid var(--primary);
-  cursor: pointer;
-}
-
-.optionsMenuCloseIcon {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  width: 4vh;
-  height: 4vh;
-  border-radius: 5px;
-}
-
-.optionsMenuCloseIcon:hover {
-  background-color: var(--primary);
-  color: black;
-}
-
-.contentContainer {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  align-content: flex-start;
-  justify-content: flex-start;
-  row-gap: 10px;
-  column-gap: 10px;
-  flex-wrap: wrap;
-  overflow: auto;
-  height: 100%;
-  max-height: 100%;
-  padding: 0.5em;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  gap: 5px;
-}
-
-.optForm.optForm {
-  flex-direction: row;
-  margin-bottom: 0;
-}
-
-.optForm.buttonRow {
-  display: flex;
-  margin-top: 0.5rem;
-  align-items: center;
-}
-
-.buttonRow {
-  margin-top: 0.25rem;
-  justify-content: center;
-  width: 90%; /* add this line */
-}
-
-.optForm.buttonDiv {
-  margin: 0 0.25em;
-}
-
-.optForm.cardDiv {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  max-height: none;
-  max-width: none;
-  height: 100%;
-}
-
-.cardList {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: flex-start;
-  gap: 1vh;
-}
-
-.checkBoxRow:last-child {
-  padding-bottom: 2vh;
-}
-
-.cardDiv {
-  flex-basis: var(--card-size);
-  max-height: var(--card-size);
-  max-width: var(--card-size);
-  height: var(--card-size);
-  width: var(--card-size);
-}
-
-@media (orientation: portrait) {
-  .optionsContainer {
-    left: 0px;
-    width: 100%;
-    max-width: 100%;
-  }
-
-  .cardDiv {
-    max-width: 14vw;
-    max-width: 14dvw;
-    max-height: 14vw;
-    max-height: 14dvw;
-  }
-}
-
-.topAndBottomContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-  height: 100%;
-  overflow-x: auto;
-}
-
-.topAndBottomText {
-  display: flex;
-  align-items: center; /* Vertically centers the text */
-  justify-content: center; /* Horizontally centers the text */
-  height: 50%;
-}
-
-.reorderCardWrapper {
-  all: unset;
-}
